backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-multipart==0.0.20
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
